# pkg/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for schedule, assessments, attendance and grades.",
        "title": "GradeFlow API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/programs": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "programs"
                ],
                "summary": "List programs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ProgramList"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "programs"
                ],
                "summary": "Create program",
                "parameters": [
                    {
                        "description": "program",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateProgram"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.Program"
                        }
                    }
                }
            }
        },
        "/programs/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "programs"
                ],
                "summary": "Get program",
                "parameters": [
                    {
                        "type": "string",
                        "description": "program id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Program"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "programs"
                ],
                "summary": "Update program",
                "parameters": [
                    {
                        "type": "string",
                        "description": "program id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "program",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateProgram"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Program"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "programs"
                ],
                "summary": "Delete program",
                "parameters": [
                    {
                        "type": "string",
                        "description": "program id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            }
        },
        "/staff": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "staff"
                ],
                "summary": "List staff",
                "parameters": [
                    {
                        "type": "string",
                        "description": "filter by department",
                        "name": "departmentId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search by name",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.StaffList"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "staff"
                ],
                "summary": "Create staff member",
                "parameters": [
                    {
                        "description": "staff",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateStaff"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.Staff"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/staff/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "staff"
                ],
                "summary": "Get staff member",
                "parameters": [
                    {
                        "type": "string",
                        "description": "staff id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Staff"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "staff"
                ],
                "summary": "Update staff member",
                "parameters": [
                    {
                        "type": "string",
                        "description": "staff id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "staff",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateStaff"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Staff"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "staff"
                ],
                "summary": "Delete staff member",
                "parameters": [
                    {
                        "type": "string",
                        "description": "staff id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.CreateProgram": {
            "type": "object",
            "required": [
                "code",
                "departmentId",
                "name"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "departmentId": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "request.CreateStaff": {
            "type": "object",
            "required": [
                "fullName"
            ],
            "properties": {
                "departmentId": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "request.UpdateProgram": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "departmentId": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "request.UpdateStaff": {
            "type": "object",
            "properties": {
                "departmentId": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "response.Error": {
            "type": "object",
            "properties": {
                "details": {},
                "error": {
                    "type": "string"
                }
            }
        },
...
        "response.Page": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "response.Program": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "departmentId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "response.ProgramList": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.Program"
                    }
                },
                "page": {
                    "$ref": "#/definitions/response.Page"
                }
            }
        },
        "response.Staff": {
            "type": "object",
            "properties": {
                "departmentId": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "response.StaffList": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.Staff"
                    }
                },
                "page": {
                    "$ref": "#/definitions/response.Page"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter the token with the `Bearer ` prefix, e.g. `Bearer \u003ctoken\u003e`",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}